+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/validation/validators/N4JSFlowgraphValidator.java
@@ -19,11 +19,11 @@ import org.eclipse.n4js.smith.DataCollector;
 import org.eclipse.n4js.smith.DataCollectors;
 import org.eclipse.n4js.smith.Measurement;
 import org.eclipse.n4js.validation.AbstractN4JSDeclarativeValidator;
-import org.eclipse.n4js.validation.flowValidators.DeadCodeValidator;
-import org.eclipse.n4js.validation.flowValidators.FlowValidator;
-import org.eclipse.n4js.validation.flowValidators.MissingReturnOrThrowValidator;
-import org.eclipse.n4js.validation.flowValidators.NullUndefinedValidator;
-import org.eclipse.n4js.validation.flowValidators.UsedBeforeDeclaredValidator;
+import org.eclipse.n4js.validation.validators.flowgraphs.DeadCodeValidator;
+import org.eclipse.n4js.validation.validators.flowgraphs.FlowValidator;
+import org.eclipse.n4js.validation.validators.flowgraphs.MissingReturnOrThrowValidator;
+import org.eclipse.n4js.validation.validators.flowgraphs.NullUndefinedValidator;
+import org.eclipse.n4js.validation.validators.flowgraphs.UsedBeforeDeclaredValidator;
 import org.eclipse.xtext.service.OperationCanceledManager;
 import org.eclipse.xtext.util.CancelIndicator;
 import org.eclipse.xtext.validation.CancelableDiagnostician;
