+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/validation/validators/flowgraphs/FlowValidator.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.validation.flowValidators;
+package org.eclipse.n4js.validation.validators.flowgraphs;
 
 import org.eclipse.n4js.flowgraphs.FlowAnalyser;
 import org.eclipse.n4js.validation.validators.N4JSFlowgraphValidator;
