+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/validation/validators/flowgraphs/UsedBeforeDeclaredValidator.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.validation.flowValidators;
+package org.eclipse.n4js.validation.validators.flowgraphs;
 
 import java.util.List;
